requirements.txt
absl-py==0.9.0
appdirs==1.4.4
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
distlib==0.3.1
et-xmlfile==1.0.1
filelock==3.0.12
gast==0.3.3
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
jdcal==1.4.1
joblib==0.16.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.0
numpy==1.18.5
oauthlib==3.1.0
openpyxl==3.0.4
opt-einsum==3.3.0
pandas==1.1.0
Pillow==7.2.0
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
six==1.15.0
sklearn-export==0.0.5
sklearn-json==0.1.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
urllib3==1.25.10
virtualenv==20.0.30
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
xlrd==1.2.0
zipp==3.1.0
pydot==1.4.1
graphviz==0.14.1
flask==1.1.2
flask_cors==3.0.8
flask_jsonpify==1.5.0
flask_restful==0.3.8